Break out format and files composables from utilities
